update to FW.F4.1.25.0

+++ f4/Middlewares/ST/STM32_USB_Host_Library/Core/Src/usbh_ioreq.c
@@ -1,31 +1,23 @@
-/** 
+/**
   ******************************************************************************
-  * @file    usbh_ioreq.c 
+  * @file    usbh_ioreq.c
   * @author  MCD Application Team
-  * @version V3.2.2
-  * @date    07-July-2015
   * @brief   This file handles the issuing of the USB transactions
   ******************************************************************************
   * @attention
   *
-  * <h2><center>&copy; COPYRIGHT 2015 STMicroelectronics</center></h2>
+  * <h2><center>&copy; Copyright (c) 2015 STMicroelectronics.
+  * All rights reserved.</center></h2>
   *
-  * Licensed under MCD-ST Liberty SW License Agreement V2, (the "License");
-  * You may not use this file except in compliance with the License.
-  * You may obtain a copy of the License at:
-  *
-  *        http://www.st.com/software_license_agreement_liberty_v2
-  *
-  * Unless required by applicable law or agreed to in writing, software 
-  * distributed under the License is distributed on an "AS IS" BASIS, 
-  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  * See the License for the specific language governing permissions and
-  * limitations under the License.
+  * This software component is licensed by ST under Ultimate Liberty license
+  * SLA0044, the "License"; You may not use this file except in compliance with
+  * the License. You may obtain a copy of the License at:
+  *                      www.st.com/SLA0044
   *
   ******************************************************************************
-  */ 
-/* Includes ------------------------------------------------------------------*/
+  */
 
+/* Includes ------------------------------------------------------------------*/
 #include "usbh_ioreq.h"
 
 /** @addtogroup USBH_LIB
@@ -35,8 +27,8 @@
 /** @addtogroup USBH_LIB_CORE
 * @{
 */
-  
-/** @defgroup USBH_IOREQ 
+
+/** @defgroup USBH_IOREQ
   * @brief This file handles the standard protocol processing (USB v2.0)
   * @{
   */
@@ -44,48 +36,48 @@
 
 /** @defgroup USBH_IOREQ_Private_Defines
   * @{
-  */ 
+  */
 
 /**
   * @}
-  */ 
- 
+  */
+
 
 /** @defgroup USBH_IOREQ_Private_TypesDefinitions
   * @{
-  */ 
+  */
 /**
   * @}
-  */ 
+  */
 
 
 
 /** @defgroup USBH_IOREQ_Private_Macros
   * @{
-  */ 
+  */
 /**
   * @}
-  */ 
+  */
 
 
 /** @defgroup USBH_IOREQ_Private_Variables
   * @{
-  */ 
+  */
 /**
   * @}
-  */ 
+  */
 /** @defgroup USBH_IOREQ_Private_FunctionPrototypes
   * @{
-  */ 
+  */
 
 /**
   * @}
-  */ 
+  */
 
 
 /** @defgroup USBH_IOREQ_Private_Functions
   * @{
-  */ 
+  */
 
 
 
@@ -97,20 +89,20 @@
   * @param  pipe_num: Pipe Number
   * @retval USBH Status
   */
-USBH_StatusTypeDef USBH_CtlSendSetup (USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
-                                uint8_t pipe_num)
+USBH_StatusTypeDef USBH_CtlSendSetup(USBH_HandleTypeDef *phost,
+                                     uint8_t *buff,
+                                     uint8_t pipe_num)
 {
 
-  USBH_LL_SubmitURB (phost,                     /* Driver handle    */
-                          pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : OUT  */
-                          USBH_EP_CONTROL,      /* EP type          */
-                          USBH_PID_SETUP,       /* Type setup       */
-                          buff,                 /* data buffer      */
-                          USBH_SETUP_PKT_SIZE,  /* data length      */ 
-                          0);
-  return USBH_OK;  
+  USBH_LL_SubmitURB(phost,                      /* Driver handle    */
+                    pipe_num,             /* Pipe index       */
+                    0U,                    /* Direction : OUT  */
+                    USBH_EP_CONTROL,      /* EP type          */
+                    USBH_PID_SETUP,       /* Type setup       */
+                    buff,                 /* data buffer      */
+                    USBH_SETUP_PKT_SIZE,  /* data length      */
+                    0U);
+  return USBH_OK;
 }
 
 
@@ -123,26 +115,26 @@ USBH_StatusTypeDef USBH_CtlSendSetup (USBH_HandleTypeDef *phost,
   * @param  pipe_num: Pipe Number
   * @retval USBH Status
   */
-USBH_StatusTypeDef USBH_CtlSendData (USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
-                                uint16_t length,
-                                uint8_t pipe_num,
-                                uint8_t do_ping )
+USBH_StatusTypeDef USBH_CtlSendData(USBH_HandleTypeDef *phost,
+                                    uint8_t *buff,
+                                    uint16_t length,
+                                    uint8_t pipe_num,
+                                    uint8_t do_ping)
 {
-  if(phost->device.speed != USBH_SPEED_HIGH)
+  if (phost->device.speed != USBH_SPEED_HIGH)
   {
-    do_ping = 0;
+    do_ping = 0U;
   }
-  
-  USBH_LL_SubmitURB (phost,                     /* Driver handle    */
-                          pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : OUT  */
-                          USBH_EP_CONTROL,      /* EP type          */                          
-                          USBH_PID_DATA,        /* Type Data        */
-                          buff,                 /* data buffer      */
-                          length,               /* data length      */ 
-                          do_ping);             /* do ping (HS Only)*/
-  
+
+  USBH_LL_SubmitURB(phost,                      /* Driver handle    */
+                    pipe_num,             /* Pipe index       */
+                    0U,                   /* Direction : OUT  */
+                    USBH_EP_CONTROL,      /* EP type          */
+                    USBH_PID_DATA,        /* Type Data        */
+                    buff,                 /* data buffer      */
+                    length,               /* data length      */
+                    do_ping);             /* do ping (HS Only)*/
+
   return USBH_OK;
 }
 
@@ -154,23 +146,23 @@ USBH_StatusTypeDef USBH_CtlSendData (USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer in which the response needs to be copied
   * @param  length: Length of the data to be received
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_CtlReceiveData(USBH_HandleTypeDef *phost, 
-                                uint8_t* buff, 
-                                uint16_t length,
-                                uint8_t pipe_num)
+USBH_StatusTypeDef USBH_CtlReceiveData(USBH_HandleTypeDef *phost,
+                                       uint8_t *buff,
+                                       uint16_t length,
+                                       uint8_t pipe_num)
 {
-  USBH_LL_SubmitURB (phost,                     /* Driver handle    */
-                          pipe_num,             /* Pipe index       */
-                          1,                    /* Direction : IN   */
-                          USBH_EP_CONTROL,      /* EP type          */                          
-                          USBH_PID_DATA,        /* Type Data        */
-                          buff,                 /* data buffer      */
-                          length,               /* data length      */ 
-                          0);
+  USBH_LL_SubmitURB(phost,                      /* Driver handle    */
+                    pipe_num,             /* Pipe index       */
+                    1U,                    /* Direction : IN   */
+                    USBH_EP_CONTROL,      /* EP type          */
+                    USBH_PID_DATA,        /* Type Data        */
+                    buff,                 /* data buffer      */
+                    length,               /* data length      */
+                    0U);
   return USBH_OK;
-  
+
 }
 
 
@@ -183,25 +175,25 @@ USBH_StatusTypeDef USBH_CtlReceiveData(USBH_HandleTypeDef *phost,
   * @param  pipe_num: Pipe Number
   * @retval USBH Status
   */
-USBH_StatusTypeDef USBH_BulkSendData (USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
-                                uint16_t length,
-                                uint8_t pipe_num,
-                                uint8_t do_ping )
-{ 
-  if(phost->device.speed != USBH_SPEED_HIGH)
+USBH_StatusTypeDef USBH_BulkSendData(USBH_HandleTypeDef *phost,
+                                     uint8_t *buff,
+                                     uint16_t length,
+                                     uint8_t pipe_num,
+                                     uint8_t do_ping)
+{
+  if (phost->device.speed != USBH_SPEED_HIGH)
   {
-    do_ping = 0;
+    do_ping = 0U;
   }
-  
-  USBH_LL_SubmitURB (phost,                     /* Driver handle    */
-                          pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : IN   */
-                          USBH_EP_BULK,         /* EP type          */                          
-                          USBH_PID_DATA,        /* Type Data        */
-                          buff,                 /* data buffer      */
-                          length,               /* data length      */  
-                          do_ping);             /* do ping (HS Only)*/
+
+  USBH_LL_SubmitURB(phost,                      /* Driver handle    */
+                    pipe_num,             /* Pipe index       */
+                    0U,                    /* Direction : IN   */
+                    USBH_EP_BULK,         /* EP type          */
+                    USBH_PID_DATA,        /* Type Data        */
+                    buff,                 /* data buffer      */
+                    length,               /* data length      */
+                    do_ping);             /* do ping (HS Only)*/
   return USBH_OK;
 }
 
@@ -213,21 +205,21 @@ USBH_StatusTypeDef USBH_BulkSendData (USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer in which the received data packet to be copied
   * @param  length: Length of the data to be received
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_BulkReceiveData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
-                                uint16_t length,
-                                uint8_t pipe_num)
+USBH_StatusTypeDef USBH_BulkReceiveData(USBH_HandleTypeDef *phost,
+                                        uint8_t *buff,
+                                        uint16_t length,
+                                        uint8_t pipe_num)
 {
-  USBH_LL_SubmitURB (phost,                     /* Driver handle    */
-                          pipe_num,             /* Pipe index       */
-                          1,                    /* Direction : IN   */
-                          USBH_EP_BULK,         /* EP type          */                          
-                          USBH_PID_DATA,        /* Type Data        */
-                          buff,                 /* data buffer      */
-                          length,               /* data length      */  
-                          0);
+  USBH_LL_SubmitURB(phost,                      /* Driver handle    */
+                    pipe_num,             /* Pipe index       */
+                    1U,                    /* Direction : IN   */
+                    USBH_EP_BULK,         /* EP type          */
+                    USBH_PID_DATA,        /* Type Data        */
+                    buff,                 /* data buffer      */
+                    length,               /* data length      */
+                    0U);
   return USBH_OK;
 }
 
@@ -239,22 +231,22 @@ USBH_StatusTypeDef USBH_BulkReceiveData(USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer in which the response needs to be copied
   * @param  length: Length of the data to be received
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_InterruptReceiveData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
-                                uint8_t length,
-                                uint8_t pipe_num)
+USBH_StatusTypeDef USBH_InterruptReceiveData(USBH_HandleTypeDef *phost,
+                                             uint8_t *buff,
+                                             uint8_t length,
+                                             uint8_t pipe_num)
 {
-  USBH_LL_SubmitURB (phost,                     /* Driver handle    */
-                          pipe_num,             /* Pipe index       */
-                          1,                    /* Direction : IN   */
-                          USBH_EP_INTERRUPT,    /* EP type          */                          
-                          USBH_PID_DATA,        /* Type Data        */
-                          buff,                 /* data buffer      */
-                          length,               /* data length      */  
-                          0); 
-  
+  USBH_LL_SubmitURB(phost,                      /* Driver handle    */
+                    pipe_num,             /* Pipe index       */
+                    1U,                   /* Direction : IN   */
+                    USBH_EP_INTERRUPT,    /* EP type          */
+                    USBH_PID_DATA,        /* Type Data        */
+                    buff,                 /* data buffer      */
+                    (uint16_t)length,     /* data length      */
+                    0U);
+
   return USBH_OK;
 }
 
@@ -265,22 +257,22 @@ USBH_StatusTypeDef USBH_InterruptReceiveData(USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer from where the data needs to be copied
   * @param  length: Length of the data to be sent
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_InterruptSendData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
-                                uint8_t length,
-                                uint8_t pipe_num)
+USBH_StatusTypeDef USBH_InterruptSendData(USBH_HandleTypeDef *phost,
+                                          uint8_t *buff,
+                                          uint8_t length,
+                                          uint8_t pipe_num)
 {
-  USBH_LL_SubmitURB (phost,                     /* Driver handle    */
-                          pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : OUT   */
-                          USBH_EP_INTERRUPT,    /* EP type          */                          
-                          USBH_PID_DATA,        /* Type Data        */
-                          buff,                 /* data buffer      */
-                          length,               /* data length      */  
-                          0);  
-  
+  USBH_LL_SubmitURB(phost,                      /* Driver handle    */
+                    pipe_num,             /* Pipe index       */
+                    0U,                   /* Direction : OUT   */
+                    USBH_EP_INTERRUPT,    /* EP type          */
+                    USBH_PID_DATA,        /* Type Data        */
+                    buff,                 /* data buffer      */
+                    (uint16_t)length,     /* data length      */
+                    0U);
+
   return USBH_OK;
 }
 
@@ -291,23 +283,23 @@ USBH_StatusTypeDef USBH_InterruptSendData(USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer in which the response needs to be copied
   * @param  length: Length of the data to be received
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_IsocReceiveData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
-                                uint32_t length,
-                                uint8_t pipe_num)
-{    
-  USBH_LL_SubmitURB (phost,                     /* Driver handle    */
-                          pipe_num,             /* Pipe index       */
-                          1,                    /* Direction : IN   */
-                          USBH_EP_ISO,          /* EP type          */                          
-                          USBH_PID_DATA,        /* Type Data        */
-                          buff,                 /* data buffer      */
-                          length,               /* data length      */
-                          0);
+USBH_StatusTypeDef USBH_IsocReceiveData(USBH_HandleTypeDef *phost,
+                                        uint8_t *buff,
+                                        uint32_t length,
+                                        uint8_t pipe_num)
+{
+  USBH_LL_SubmitURB(phost,                      /* Driver handle    */
+                    pipe_num,             /* Pipe index       */
+                    1U,                   /* Direction : IN   */
+                    USBH_EP_ISO,          /* EP type          */
+                    USBH_PID_DATA,        /* Type Data        */
+                    buff,                 /* data buffer      */
+                    (uint16_t)length,     /* data length      */
+                    0U);
 
-  
+
   return USBH_OK;
 }
 
@@ -318,31 +310,27 @@ USBH_StatusTypeDef USBH_IsocReceiveData(USBH_HandleTypeDef *phost,
   * @param  buff: Buffer pointer from where the data needs to be copied
   * @param  length: Length of the data to be sent
   * @param  pipe_num: Pipe Number
-  * @retval USBH Status. 
+  * @retval USBH Status.
   */
-USBH_StatusTypeDef USBH_IsocSendData(USBH_HandleTypeDef *phost, 
-                                uint8_t *buff, 
-                                uint32_t length,
-                                uint8_t pipe_num)
+USBH_StatusTypeDef USBH_IsocSendData(USBH_HandleTypeDef *phost,
+                                     uint8_t *buff,
+                                     uint32_t length,
+                                     uint8_t pipe_num)
 {
-  USBH_LL_SubmitURB (phost,                     /* Driver handle    */
-                          pipe_num,             /* Pipe index       */
-                          0,                    /* Direction : OUT   */
-                          USBH_EP_ISO,          /* EP type          */                          
-                          USBH_PID_DATA,        /* Type Data        */
-                          buff,                 /* data buffer      */
-                          length,               /* data length      */ 
-                          0);
-  
+  USBH_LL_SubmitURB(phost,                      /* Driver handle    */
+                    pipe_num,             /* Pipe index       */
+                    0U,                   /* Direction : OUT   */
+                    USBH_EP_ISO,          /* EP type          */
+                    USBH_PID_DATA,        /* Type Data        */
+                    buff,                 /* data buffer      */
+                    (uint16_t)length,     /* data length      */
+                    0U);
+
   return USBH_OK;
 }
 /**
 * @}
-*/ 
-
-/**
-* @}
-*/ 
+*/
 
 /**
 * @}
@@ -350,7 +338,11 @@ USBH_StatusTypeDef USBH_IsocSendData(USBH_HandleTypeDef *phost,
 
 /**
 * @}
-*/ 
+*/
+
+/**
+* @}
+*/
 
 /************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
